move tests to be inline
